requirements.txt
esm
biopython
torch
scikit-learn
numpy
